chore: extract aggregerEvolution shared util
